+++ test/integ/test_const.py
@@ -1,0 +1,18 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+
+"""
+Maya Deadline Cloud - Constants for file name and folder to use for testing
+"""
+
+# File name
+TEMPLATE = "template.yaml"
+ASSET_REFERENCES = "asset_references.yaml"
+PARAMETER_VALUES = "parameter_values.yaml"
+STICKY_SETTING = "deadline_render_settings.json"
+
+# Folder name
+TEST_SCENE_FOLDER = "scene"
+EXPECTED_OUTPUT_FOLDER = "expected_images"
+EXPECTED_JOB_BUNDLE_FOLDER = "expected_job_bundle"
+OUTPUT_FOLDER = "output"
+JOB_HISTORY_FOLDER = "jobhistory"